+++ src/web/command_control/package.json
@@ -55,7 +55,7 @@
     "html-webpack-plugin": "^4.5.2",
     "idb": "^7.0.2",
     "json-loader": "^0.5.7",
-    "jszip": "*3.10.1",
+    "jszip": "^2.2.0",
     "less": "^3.13.1",
     "less-loader": "^11.1.0",
     "less-vars-loader": "^1.1.0",
